+++ internal/adf/update_nonwindows.go
@@ -1,0 +1,5 @@
+//go:build !windows
+
+package adf
+
+func startDetachedWindows(string) error { return nil }